Sprint3 Initial Commit

+++ POS/src/main/java/com/ust/pos/customer/service/impl/CustomerServiceImpl.java
@@ -4,6 +4,7 @@ import com.ust.pos.customer.service.AddressService;
 import com.ust.pos.customer.service.CustomerService;
 import com.ust.pos.dto.AddressDto;
 import com.ust.pos.dto.CustomerDto;
+import com.ust.pos.dto.WsDto;
 import com.ust.pos.model.Customer;
 import com.ust.pos.model.CustomerRepository;
 import org.modelmapper.ModelMapper;
@@ -31,18 +32,23 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public CustomerDto findByIdentifier(String identifier) {
+
         Customer customer = customerRepository.findByIdentifier(identifier);
         if (customer == null) {
             return null;
         }
+
         return modelMapper.map(customer, CustomerDto.class);
     }
 
     @Override
     public CustomerDto save(CustomerDto customerDto) {
+
         String identifier = customerDto.getIdentifier();
         Customer existingCustomer = customerRepository.findByIdentifier(identifier);
+
         if (existingCustomer != null) {
+
             customerDto.setMessage("Customer with identifier - " + identifier + " already exists");
             customerDto.setSuccess(false);
             return customerDto;
@@ -90,6 +96,7 @@ public class CustomerServiceImpl implements CustomerService {
                 findByPhoneNoAndAddressType(existingCustomer.getPhoneNo(), "billingAddress"));
         customerDto.setShippingAddress(addressService.
                 findByPhoneNoAndAddressType(existingCustomer.getPhoneNo(), "shippingAddress"));
+
         customerRepository.save(existingCustomer);
         return customerDto;
     }
@@ -98,16 +105,37 @@ public class CustomerServiceImpl implements CustomerService {
     @Transactional
     public void delete(String identifier) {
         Customer customer = customerRepository.findByIdentifier(identifier);
-        addressService.deleteByPhoneNo(customer.getPhoneNo());
+
+        if (customer != null) {
+            addressService.deleteByPhoneNo(customer.getPhoneNo());
+        }
+
         customerRepository.deleteByIdentifier(identifier);
     }
 
     @Override
-    public List<CustomerDto> findAll(Pageable pageable) {
+    public WsDto<CustomerDto> findAll(Pageable pageable) {
+
         Type listType = new TypeToken<List<CustomerDto>>() {
         }.getType();
+
         Page<Customer> customerPage = customerRepository.findAll(pageable);
-        return modelMapper.map(customerPage.getContent(), listType);
+
+        List<CustomerDto> customerDtos = modelMapper.map(
+                customerPage.getContent(),
+                listType
+        );
+
+        WsDto<CustomerDto> wsDto =
+                new WsDto<>();
+
+        wsDto.setContent(customerDtos);
+        wsDto.setPage(customerPage.getNumber());
+        wsDto.setSizePerPage(customerPage.getSize());
+        wsDto.setTotalPages(customerPage.getTotalPages());
+        wsDto.setTotalRecords(customerPage.getTotalElements());
+
+        return wsDto;
     }
 
     @Override
@@ -118,4 +146,5 @@ public class CustomerServiceImpl implements CustomerService {
             customerRepository.save(customer);
         }
     }
+
 }
